+++ src/main/java/com/dominik/typer/model/entity/BetEntity.java
@@ -7,8 +7,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @ToString
@@ -24,6 +22,6 @@ public class BetEntity {
     Integer userId;
     Integer pointsFromBetEvent;
     BetType betType;
-    BigDecimal betAmount;
+    Integer betAmount;
 
 }
